requirements/base.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-phonenumber-field==6.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
Faker==8.12.1
Pillow==8.3.1
# psycopg2==2.8.5
psycopg2-binary==2.8.4
sqlparse==0.4.2
backports.zoneinfo==0.2.1;python_version<"3.9"
pytz==2022.5
